Refactor AuthenticationClient setup
Also refactor naming to be more specific: AppSettings -> OidcSettings
