+++ hestia/hestia/Models/OidcSettings.cs
@@ -4,7 +4,7 @@ namespace hestia.Models
     /// <summary>
     /// Environment variables model following appSettings.json structure.
     /// </summary>
-	public class AppSettings
+	public class OidcSettings
 	{
 		public string AuthenticationDomain { get; set; }
         public string ClientId { get; set; }
