+++ hestia/hestia/appSettings.json.template
@@ -1,5 +1,5 @@
 {
-  "AppSettings": {
+  "OidcSettings": {
     "AuthenticationDomain": "",
     "ClientId": "",
     "ClientSecret": "",
